Alter existing cast function for eql and remove newly introduced
