+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/EqlQueryRenderer.java
@@ -1009,8 +1009,8 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 			builder.append(visit(ctx.case_expression()));
 		} else if (ctx.entity_type_expression() != null) {
 			builder.append(visit(ctx.entity_type_expression()));
-		} else if (ctx.cast_expression() != null) {
-			return (visit(ctx.cast_expression()));
+		} else if (ctx.cast_function() != null) {
+			return (visit(ctx.cast_function()));
 		}
 
 		return builder;
@@ -1935,6 +1935,7 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 			builder.appendInline(visit(ctx.string_expression(0)));
 			builder.append(TOKEN_CLOSE_PAREN);
 		} else if (ctx.LEFT() != null) {
+
 			builder.append(QueryTokens.token(ctx.LEFT()));
 			builder.append(TOKEN_OPEN_PAREN);
 			builder.appendInline(visit(ctx.string_expression(0)));
@@ -1942,6 +1943,7 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 			builder.appendInline(visit(ctx.arithmetic_expression(0)));
 			builder.append(TOKEN_CLOSE_PAREN);
 		} else if (ctx.RIGHT() != null) {
+
 			builder.append(QueryTokens.token(ctx.RIGHT()));
 			builder.append(TOKEN_OPEN_PAREN);
 			builder.appendInline(visit(ctx.string_expression(0)));
@@ -1949,6 +1951,7 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 			builder.appendInline(visit(ctx.arithmetic_expression(0)));
 			builder.append(TOKEN_CLOSE_PAREN);
 		} else if (ctx.REPLACE() != null) {
+
 			builder.append(QueryTokens.token(ctx.REPLACE()));
 			builder.append(TOKEN_OPEN_PAREN);
 			builder.appendInline(visit(ctx.string_expression(0)));
@@ -1983,9 +1986,9 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 		builder.append(TOKEN_OPEN_PAREN);
 		builder.appendInline(visit(ctx.single_valued_path_expression()));
 		builder.append(TOKEN_SPACE);
-		builder.appendInline(visit(ctx.identification_variable()));
+		builder.appendInline(QueryTokenStream.concat(ctx.identification_variable(), this::visit, TOKEN_SPACE));
 
-		if (ctx.numeric_literal() != null) {
+		if (!ObjectUtils.isEmpty(ctx.numeric_literal())) {
 
 			builder.append(TOKEN_OPEN_PAREN);
 			builder.appendInline(QueryTokenStream.concat(ctx.numeric_literal(), this::visit, TOKEN_COMMA));
@@ -2084,18 +2087,6 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 		} else {
 			return visit(ctx.nullif_expression());
 		}
-	}
-
-	@Override
-	public QueryRendererBuilder visitCast_expression(EqlParser.Cast_expressionContext ctx) {
-		QueryRendererBuilder builder = QueryRenderer.builder();
-		builder.append(QueryTokens.token(ctx.CAST()));
-		builder.append(TOKEN_OPEN_PAREN);
-		builder.appendInline(visit(ctx.string_expression()));
-		builder.append(QueryTokens.expression(ctx.AS()));
-		builder.appendInline(visit(ctx.type_literal()));
-		builder.append(TOKEN_CLOSE_PAREN);
-		return builder;
 	}
 
 	@Override
@@ -2226,9 +2217,11 @@ class EqlQueryRenderer extends EqlBaseVisitor<QueryTokenStream> {
 			return QueryRendererBuilder.from(QueryTokens.expression(ctx.IDENTIFICATION_VARIABLE()));
 		} else if (ctx.f != null) {
 			return QueryRendererBuilder.from(QueryTokens.expression(ctx.f));
-		} else {
-			return QueryRenderer.builder();
+		} else if (ctx.type_literal() != null) {
+			return visit(ctx.type_literal());
 		}
+
+		return QueryRenderer.builder();
 	}
 
 	@Override
